* Added Maven * Fixed formatting * Renamed VersionFormater to VersionFormatter (typo) * Improved VersionFormatter
